resolve remarks to Restructuring proposal #39
All remarks issued by fbtom incorporated in commit.
