+++ Dockerfile
@@ -11,7 +11,7 @@ COPY . /application
 
 RUN cmake -S . -B build && cmake --build build
 
-RUN /application/bin/${PROJECT_NAME}-all-tests
+RUN /application/bin/${PROJECT_NAME}-tests
 
 FROM ubuntu:mantic as FINAL
 ENV PROJECT_NAME=desktop-business-app
@@ -22,6 +22,6 @@ RUN apt update -y && \
 
 WORKDIR /application
 COPY --from=BUILD /application/bin/${PROJECT_NAME} /application/bin/${PROJECT_NAME}
-COPY --from=BUILD /application/bin/${PROJECT_NAME}-all-tests /application/bin/${PROJECT_NAME}-all-tests
+COPY --from=BUILD /application/bin/${PROJECT_NAME}-tests /application/bin/${PROJECT_NAME}-tests
 
 ENTRYPOINT /application/bin/${PROJECT_NAME}
