+++ README.md
@@ -24,7 +24,7 @@ docker run -it --entrypoint /bin/bash desktop-business-application
 
 ### Run all tests from image
 ```bash
-docker run -it --entrypoint="/application/bin/desktop-business-app-all-tests" desktop-business-application
+docker run -it --entrypoint="/application/bin/desktop-business-app-tests" desktop-business-application
 ```
 
 ### Run GUI application from docker (on Ubuntu)
