+++ src/tests/ut/receptionist/shift_test.cpp
@@ -1,7 +1,6 @@
 #include "shift_test.hpp"
-#include "shift.hpp"
 
-namespace unit_test::receptionist
+namespace
 {
 
 TEST_P(ShiftParameterizedFixture, GivenShiftEnumerationExpectValidConversionToString)
@@ -15,4 +14,4 @@ INSTANTIATE_TEST_SUITE_P(ShiftConversionTest, ShiftParameterizedFixture,
                          ::testing::Values(TestParams{Shift::Morning, "morning shift"},
                                            TestParams{Shift::Afternoon, "afternoon shift"}));
 
-} // namespace unit_test::receptionist
+} // namespace
