+++ src/tests/ut/receptionist/shift_test.hpp
@@ -3,9 +3,6 @@
 #include "shift.hpp"
 #include "gtest/gtest.h"
 
-namespace unit_test::receptionist
-{
-
 struct TestParams
 {
     Shift shift_input_;
@@ -15,5 +12,3 @@ struct TestParams
 class ShiftParameterizedFixture : public ::testing::TestWithParam<TestParams>
 {
 };
-
-} // namespace unit_test::receptionist
